refactor(transform): move api classes to separate file
